divide io test
